refactor: move `addDevRules` to css class feature
