+++ packages/core/src/helpers/rule.ts
@@ -34,45 +34,6 @@ export function isInConditionalGroup(node: postcss.Rule | postcss.AtRule, includ
     );
 }
 
-export function createWarningRule(
-    extendedNode: string,
-    scopedExtendedNode: string,
-    extendedFile: string,
-    extendingNode: string,
-    scopedExtendingNode: string,
-    extendingFile: string,
-    useScoped = false
-) {
-    const message = `"class extending component '.${extendingNode} => ${scopedExtendingNode}' in stylesheet '${extendingFile}' was set on a node that does not extend '.${extendedNode} => ${scopedExtendedNode}' from stylesheet '${extendedFile}'" !important`;
-    return postcss.rule({
-        selector: `.${useScoped ? scopedExtendingNode : extendingNode}:not(.${
-            useScoped ? scopedExtendedNode : extendedNode
-        })::before`,
-        nodes: [
-            postcss.decl({
-                prop: 'content',
-                value: message,
-            }),
-            postcss.decl({
-                prop: 'display',
-                value: `block !important`,
-            }),
-            postcss.decl({
-                prop: 'font-family',
-                value: `monospace !important`,
-            }),
-            postcss.decl({
-                prop: 'background-color',
-                value: `red !important`,
-            }),
-            postcss.decl({
-                prop: 'color',
-                value: `white !important`,
-            }),
-        ],
-    });
-}
-
 export function createSubsetAst<T extends postcss.Root | postcss.AtRule>(
     root: postcss.Root | postcss.AtRule,
     selectorPrefix: string,
